+++ src/esm/components/form/select/dropdown-select/combobox.js
@@ -218,7 +218,9 @@ export var DropdownCombobox = function DropdownCombobox(_ref) {
     "aria-label": a11yStatusValid
   }))), /*#__PURE__*/React.createElement("ul", _extends({
     className: "k-Form-Dropdown__list"
-  }, getMenuProps()), isOpen && (filteredOptions.length > 0 ? filteredOptions.map(function (item, index) {
+  }, getMenuProps(), {
+    role: isOpen ? 'listbox' : null
+  }), isOpen && (filteredOptions.length > 0 ? filteredOptions.map(function (item, index) {
     return /*#__PURE__*/React.createElement("li", _extends({
       className: classNames('k-Form-Dropdown__item', "k-Form-Dropdown__item--level_" + (item.level || 1), {
         'k-Form-Dropdown__item--higlighted': highlightedIndex === index