+++ src/esm/components/structure/cards/horizontal-crowdfunding-card/index.js
@@ -1,6 +1,6 @@
 import _extends from "@babel/runtime/helpers/extends";
 import _objectWithoutPropertiesLoose from "@babel/runtime/helpers/objectWithoutPropertiesLoose";
-var _excluded = ["href", "imageProps", "title", "titleTag", "description", "info", "progress", "progressColor", "className", "noProgressOnMobile", "noProgress"];
+var _excluded = ["href", "imageProps", "title", "titleTag", "description", "info", "progress", "progressColor", "className", "noProgressOnMobile", "noProgress", "progressLabel"];
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
@@ -39,6 +39,8 @@ export var HorizontalCrowdfundingCard = function HorizontalCrowdfundingCard(_ref
       noProgressOnMobile = _ref$noProgressOnMobi === void 0 ? false : _ref$noProgressOnMobi,
       _ref$noProgress = _ref.noProgress,
       noProgress = _ref$noProgress === void 0 ? false : _ref$noProgress,
+      _ref$progressLabel = _ref.progressLabel,
+      progressLabel = _ref$progressLabel === void 0 ? 'progression' : _ref$progressLabel,
       props = _objectWithoutPropertiesLoose(_ref, _excluded);
 
   return /*#__PURE__*/React.createElement(StyledCard, _extends({
@@ -65,6 +67,7 @@ export var HorizontalCrowdfundingCard = function HorizontalCrowdfundingCard(_ref
   }, info), !noProgress && /*#__PURE__*/React.createElement("div", {
     className: "k-HorizontalCrowdfundingCard__progress"
   }, /*#__PURE__*/React.createElement(Progress, {
+    "aria-label": progressLabel,
     value: progress,
     color: progressColor,
     rampProps: {
@@ -88,6 +91,7 @@ HorizontalCrowdfundingCard.propTypes = {
   info: PropTypes.node,
   progress: PropTypes.number,
   progressColor: PropTypes.string,
+  progressLabel: PropTypes.string,
   noProgress: PropTypes.bool,
   noProgressOnMobile: PropTypes.bool
 };